fix(suite): return sorted data from getAccountTransactions util
- move getAccountTransactions from account to transaction utils
- use `account.key` as parameter
- sort result by blockHeight
